updateBrakePower: test

+++ ISO 19030/createSFOCCoefficients.sql
@@ -1,3 +1,3 @@
 /* Create table for coefficients of SFOC fitted curve */
 
-CREATE TABLE SFOCCoefficients (id INT PRIMARY KEY AUTO_INCREMENT, Engine_Model VARCHAR(100), XO DOUBLE(10, 5), X1 DOUBLE(10, 5),  X2 DOUBLE(10, 5));
+CREATE TABLE SFOCCoefficients (id INT PRIMARY KEY AUTO_INCREMENT, Engine_Model VARCHAR(100), XO DOUBLE(20, 10), X1 DOUBLE(20, 10),  X2 DOUBLE(20, 10));
